+++ hi_scripting/scripting/scriptnode/api/StaticNodeWrappers.cpp
@@ -143,7 +143,7 @@ juce::Rectangle<int> WrapperNode::createRectangleForParameterSliders(int numColu
 	else
 	{
 		int numParameters = getNumParameters();
-		int numRows = (int)std::ceilf((float)numParameters / (float)numColumns);
+		int numRows = (int)std::ceil((float)numParameters / (float)numColumns);
 
 		h += numRows * (48 + 18);
 		w = jmin(numColumns * 100, numParameters * 100);
